Seed TypeScript plugin for VSIX smoke
