+++ scripts/smoke-installed-vsix.mjs
@@ -2,7 +2,7 @@ import { existsSync, mkdtempSync, readFileSync, readdirSync, rmSync } from 'node
 import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { createRequire } from 'node:module';
-import { fileURLToPath } from 'node:url';
+import { fileURLToPath, pathToFileURL } from 'node:url';
 
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
 const requireFromExtension = createRequire(
@@ -56,6 +56,7 @@ function findVsixForTarget({ artifactsDir, version, target }) {
 async function smokeInstalledVsix({ target, vsixPath }) {
   const profilePath = mkdtempSync(path.join(tmpdir(), 'cg-vsix-'));
   const harnessPath = path.join(profilePath, 'harness');
+  const homeDir = path.join(profilePath, 'home');
   const userDataDir = path.join(profilePath, 'user-data');
   const extensionsDir = path.join(profilePath, 'extensions');
   const workspacePath = path.join(repoRoot, 'examples', 'example-typescript');
@@ -77,6 +78,13 @@ async function smokeInstalledVsix({ target, vsixPath }) {
 
   try {
     await writeHarnessExtension(harnessPath);
+    const { e2eScenarios, prepareScenarioWorkspacePlugins } = await loadE2ESmokeSetup();
+    const scenario = e2eScenarios.find(entry => entry.name === 'typescript');
+    if (!scenario) {
+      throw new Error('Missing TypeScript E2E scenario');
+    }
+    prepareScenarioWorkspacePlugins(scenario, repoRoot, workspacePath, homeDir, false);
+
     await runVSCodeCommand([
       ...profileArgs,
       '--install-extension',
@@ -90,6 +98,7 @@ async function smokeInstalledVsix({ target, vsixPath }) {
       extensionTestsEnv: {
         CODEGRAPHY_E2E_SCENARIO: 'typescript',
         CODEGRAPHY_E2E_GREP: 'extension activates without error|all commands are registered|manual graph indexing creates scenario edges',
+        HOME: homeDir,
       },
       launchArgs: [
         workspacePath,
@@ -109,6 +118,34 @@ async function smokeInstalledVsix({ target, vsixPath }) {
   } finally {
     rmSync(profilePath, { recursive: true, force: true });
   }
+}
+
+async function loadE2ESmokeSetup() {
+  const scenariosModule = await import(pathToFileURL(path.join(
+    repoRoot,
+    'packages',
+    'extension',
+    'dist-e2e',
+    'extension',
+    'src',
+    'e2e',
+    'scenarios.js',
+  )).href);
+  const workspacePluginsModule = await import(pathToFileURL(path.join(
+    repoRoot,
+    'packages',
+    'extension',
+    'dist-e2e',
+    'extension',
+    'src',
+    'e2e',
+    'workspacePlugins.js',
+  )).href);
+
+  return {
+    e2eScenarios: scenariosModule.e2eScenarios,
+    prepareScenarioWorkspacePlugins: workspacePluginsModule.prepareScenarioWorkspacePlugins,
+  };
 }
 
 async function writeHarnessExtension(harnessPath) {
